consistency_check/requirements.txt
apscheduler
connexion[flask,uvicorn,swagger-ui]
httpx
pykafka
PyYAML
SQLAlchemy
mysqlclient
setuptools